requirements.txt
pyinstaller
pywebview
fastapi
uvicorn
sqlalchemy
alembic
pydantic
deepdiff
Faker
pytest-benchmark
httpx